Get data from server and show in home page

+++ data/src/main/java/xyz/zohre/data/model/Products.kt
@@ -9,7 +9,7 @@ data class Products (
     @Json(name = "reviewInformation") val reviewInformation : ReviewInformation,
     @Json(name = "USPs") val uSPs : List<String>,
     @Json(name = "availabilityState") val availabilityState : Int,
-    @Json(name = "salesPriceIncVat") val salesPriceIncVat : Int,
+    @Json(name = "salesPriceIncVat") val salesPriceIncVat : Double,
     @Json(name = "productImage") val productImage : String,
     @Json(name = "nextDayDelivery") val nextDayDelivery : Boolean
 )
